+++ optalg/opt_solver/ipopt.py
@@ -1,7 +1,7 @@
 #****************************************************#
 # This file is part of OPTALG.                       #
 #                                                    #
-# Copyright (c) 2015, Tomas Tinoco De Rubira.        #
+# Copyright (c) 2019, Tomas Tinoco De Rubira.        #
 #                                                    #
 # OPTALG is released under the BSD 2-clause license. #
 #****************************************************#
@@ -9,7 +9,7 @@
 from __future__ import print_function
 import numpy as np
 from .opt_solver_error import *
-from .problem import cast_problem
+from .problem import cast_problem, OptProblem
 from .opt_solver import OptSolver
 from scipy.sparse import bmat
 
@@ -37,6 +37,18 @@ class OptSolverIpopt(OptSolver):
         
         OptSolver.__init__(self)
         self.parameters = OptSolverIpopt.parameters.copy()
+
+    def supports_properties(self, properties):
+
+        for p in properties:
+            if p not in [OptProblem.PROP_CURV_LINEAR,
+                         OptProblem.PROP_CURV_QUADRATIC,
+                         OptProblem.PROP_CURV_NONLINEAR,
+                         OptProblem.PROP_VAR_CONTINUOUS,
+                         OptProblem.PROP_TYPE_FEASIBILITY,
+                         OptProblem.PROP_TYPE_OPTIMIZATION]:
+                return False
+        return True
 
     def create_ipopt_context(self):
         
@@ -96,7 +108,7 @@ class OptSolverIpopt(OptSolver):
                             eval_jac_g,
                             eval_h)
                 
-    def solve(self,problem):
+    def solve(self, problem):
         
         # Local vars
         params = self.parameters